+++ example/src/main/java/io/realm/examples/adapters/ui/listview/ListViewExampleActivity.java
@@ -47,7 +47,7 @@ public class ListViewExampleActivity extends AppCompatActivity {
         // RealmResults are "live" views, that are automatically kept up to date, even when changes happen
         // on a background thread. The RealmBaseAdapter will automatically keep track of changes and will
         // automatically refresh when a change is detected.
-        RealmResults<Item> counters = realm.where(Item.class).findAllSorted(Item.FIELD_ID);
+        RealmResults<Item> counters = realm.where(Item.class).sort(Item.FIELD_ID).findAll();
         adapter = new MyListAdapter(counters);
 
         ListView listView = (ListView) findViewById(R.id.listView);